refactor: Add config. Flatten project structure
Bump README.md for new release (0.0.3)
