+++ SilenceTameWolfCubMod.csproj
