+++ SilenceTameWolfCubMod.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.33027.164
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SilenceTameWolfCubMod", "SilenceTameWolfCubMod\SilenceTameWolfCubMod.csproj", "{9E647534-F920-4303-98C2-9574A5845DEE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SilenceTameWolfCubMod", "SilenceTameWolfCubMod.csproj", "{9E647534-F920-4303-98C2-9574A5845DEE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
